+++ upload-cli/package.json
@@ -8,8 +8,12 @@
   "dependencies": {
     "@google-cloud/storage": "^4.5.0",
     "dotenv": "^8.2.0",
+    "inquirer": "^7.0.6",
     "shared": "1.0.0",
     "sharp": "^0.25.0",
     "ts-exif-parser": "^0.1.5"
+  },
+  "devDependencies": {
+    "@types/inquirer": "^6.5.0"
   }
 }
