+++ upload-cli/readme.md
@@ -1,0 +1,21 @@
+Martin Blanco API Upload CLI
+
+Source directory as an absolute path, e.g.: ~/Desktop/upload:
+
+Target environment:
+
+1. Development
+2. Production
+
+Get files from path:
+
+Get file name:
+
+1 of 100
+
+Resizing and rotating image
+Getting target file path
+Delay of ms
+Uploading file
+
+Finish.
